Implements a twitter-based newsfeed page that passes all the basic tests as well as passes the advamced requirements.
Adds a three-column display for viewing a user's twiddler timeline, a
main twiddler newsfeed, and an options tab for refreshing the twiddler
feed and adding a new twiddle. Each twiddle is displayed with a random
profile picture, a username, a user twiddler handle, a twiddler message,
and a automatically generated created at date.